% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiplot.R
\name{asciiplot}
\alias{asciiplot}
\alias{print.asciiplot}
\title{ASCII plots}
\usage{
asciiplot(df, aes = NULL, geom = NULL)

\method{print}{asciiplot}(x, ...)
}
\arguments{
\item{df}{a data frame, ideally two continuous variables and a factor}

\item{aes}{a list of aesthetics \code{x, y} and (optionally) \code{shape}}

\item{geom}{initial geometry; only \code{"point"} so far implemented}

\item{x}{an \code{asciiplot} object}

\item{...}{not used}
}
\value{
an \code{asciiplot} object, which can be \code{print}ed
}
\description{
Create a \emph{Grammar of Graphics}-style visualisation without a graphical device!
  Simply \code{print} your results to the console as beautiful data-driven ASCII art. The
  renderer will automatically scale to the width of your R console. Currently there are
  no borders, titles, ticks, labels or legends; these may or may not be added in future.
}
\details{
The rendering height is fixed to 25 characters by default, but can be tweaked
 via \code{ap$height <- newheight} where \code{ap} is your saved \code{asciiplot} object.
}
\examples{
asciiplot(iris, aes=list(x='Sepal.Width', y='Sepal.Length', shape='Species'), geom='point')

if(!requireNamespace('BradleyTerry2')) install.packages('BradleyTerry2')
asciiplot(ggplot2::diamonds, aes=list(x='carat', y='price', shape='cut'), geom='point')

}
\references{
Wilkinson, L. (2006).
\emph{The Grammar of Graphics.}
Springer Science & Business Media.

Wickham, H. (2010).
A layered grammar of graphics.
\emph{Journal of Computational and Graphical Statistics}
19(1), 3--28.
}
\seealso{
Other plot constructors: \code{\link{linearScale}},
  \code{\link{logScale}}
}
\concept{
ascii textplot utf8plot unicodeplot
}

