% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BradleyTerry.R
\name{BT2}
\alias{BT2}
\title{Estimate the parameters of a Bradley--Terry model}
\usage{
BT2(C, sort = FALSE)
}
\arguments{
\item{C}{a square matrix of paired comparison data}

\item{sort}{logical. If \code{TRUE}, sort the weights in descending order}
}
\description{
A one-liner to extract model parameters from \code{BradleyTerry2}.
Deprecated: use \code{\link{BradleyTerry}} or \code{\link{BTscores}} instead.
}
\details{
Previously a placeholder function, now remains for reference and comparison purposes.
}
\examples{
BT2(citations)

# Compare:
plot(BT2(citations), BTscores(citations))
abline(0, 1)

}
\references{
Heather Turner, David Firth (2012).
Bradley--Terry Models in R: The \code{BradleyTerry2} Package.
\emph{Journal of Statistical Software},
48(9), 1--21.

Bradley, R. A., & Terry, M. E. (1952).
Rank analysis of incomplete block designs: I. The method of paired comparisons.
\emph{Biometrika},
39(3/4), 324--345.
}
\seealso{
BradleyTerry BTscores
}

