% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BradleyTerry.R
\name{BTscores}
\alias{BTscores}
\title{Estimate the parameters of a Bradley--Terry model}
\usage{
BTscores(X, sort = FALSE)
}
\arguments{
\item{X}{a matrix of paired comparison data}

\item{sort}{logical.}
}
\description{
A quick way to retrieve Bradley--Terry parameter estimates without saving the \code{glm} object.
}
\details{
For a model summary and standard errors, use \code{\link[=BradleyTerry]{BradleyTerry()}}.
}
\examples{
BTscores(citations)

}
\seealso{
Other network centrality estimators: \code{\link{BradleyTerry}},
  \code{\link{ILSR}}, \code{\link{PageRank}},
  \code{\link{Scroogefactor}}
}
