% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation-data.R
\docType{data}
\name{citations}
\alias{citations}
\title{Citation data}
\format{An object of class \code{table} with 4 rows and 4 columns.}
\usage{
data(citations)
}
\description{
Taken from the \code{BradleyTerry2} package.
}
\details{
A 4-by-4 table of citation counts between statistical journals.
}
\keyword{datasets}

