% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{community_profile}
\alias{community_profile}
\title{Calculate community profiles}
\usage{
community_profile(network, ..., self = TRUE)
}
\arguments{
\item{network}{A matrix of citation counts (from columns to rows) or an \code{\link[igraph]{igraph}} object.}

\item{\dots}{Further arguments passed to \code{\link{merge_communities}}.}

\item{self}{logical. Include self-citations? If \code{FALSE}, they will not be counted.}
}
\value{
A sparse matrix of community profiles.
Each column corresponds to a community and each row to a node in the network.
}
\description{
Given a directed network and a mapping of nodes to communities, calculate the \emph{community profiles}
--- the vector of probabilities that a node from one community cites other nodes in the network.
}
\details{
Profiles are computed by aggregating the out-links of nodes in each community, then scaling to sum
to one. This means nodes with more out-links have greater influence on their community profiles.
}
\seealso{
\code{\link{merge_communities}}, \code{\link{cprofile}}
}
