% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_point.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Cosine similarity between two vectors}
\usage{
cosine_similarity(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector, the same length as \code{x}}
}
\value{
Generally, a scalar between \eqn{-1} and \eqn{1}.
Or, if \code{x} and \code{y} are non-negative, a value between \eqn{0} and \eqn{1}.
}
\description{
This function computes the \emph{cosine similarity} between two vectors, defined by
\deqn{\cos(\theta) = \frac{x \cdot y}{\|x\|_2\|y\|_2}.}{cos(\theta) = (x . y) / (|x| |y|).}
}
\details{
Cosine similarity is related to Euclidean distance by
\deqn{\|x - y \|^2 = 2(1 - \cos(\theta)).}{|x - y|^2 = 2(1 - cos(\theta)).}
So \deqn{\cos(\theta) = 1 - \frac12\|x - y\|^2,}{cos(\theta) = 1 - 0.5 * |x - y|^2,}
assuming \code{x} and \code{y} have been normalised to be unit vectors.
Therefore, if we want to \emph{maximise} cosine similarity, we can minimise Euclidean distance
and then make the conversion. See \code{\link[=nearest_point]{nearest_point()}}.
}
\references{
\link{https://en.wikipedia.org/wiki/Cosine_similarity}
}
\seealso{
\code{\link[=nearest_cosine]{nearest_cosine()}}, \code{\link[=cos]{cos()}}, \code{\link[=acos]{acos()}}
}
