% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiles.R
\name{cprofile}
\alias{cprofile}
\alias{cprofile.default}
\alias{cprofile.igraph}
\title{Calculate citation profiles}
\usage{
cprofile(x, self)

\method{cprofile}{default}(x, self = TRUE)

\method{cprofile}{igraph}(x, self = TRUE)
}
\arguments{
\item{x}{a matrix of citations (from columns to rows) or an \link[igraph:igraph]{igraph} object}

\item{self}{logical. Include self-citations? If \code{FALSE}, they will not be counted.}
}
\value{
A matrix of journal citation profiles.
}
\description{
For a directed network, calculate the outgoing transition probabilities of a random walker at each node.
This is equivalent to the proportion of citations each journal allocates to each other journal.
}
\examples{
cprofile(citations)
cprofile(citations, self = FALSE)

}
\seealso{
\code{\link[=community_profile]{community_profile()}}
}
