% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_point.R
\name{dissimilarity}
\alias{dissimilarity}
\title{Index of dissimilarity}
\usage{
dissimilarity(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{y}{A numeric vector, the same length as \code{x}}
}
\value{
A number between \eqn{0} and \eqn{1}. Lower is better.
}
\description{
Index of dissimilarity
}
\references{
Kuha, Jouni and Firth, David (2011).
On index of dissimilarity for lack of fit in loglinear and log-multiplicative models.
\emph{Computational Statistics & Data Analysis}, 55(1):375--388.
}
\seealso{
\code{\link[=profile_residuals]{profile_residuals()}}
}
