% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{fitted_citations}
\alias{fitted_citations}
\title{Predict citations based on community profile}
\usage{
fitted_citations(idx = NULL, citations, communities, self = TRUE)
}
\arguments{
\item{idx}{A journal name or index. Vectorised for \code{nearest_profile()}.}

\item{communities}{A membership vector or \link[igraph:communities]{igraph::communities} object.}

\item{self}{logical. Include self-citations? If \code{FALSE}, they will not be counted.}
}
\value{
A vector of predicted outgoing citations from journal \code{idx}.
}
\description{
Based on the nearest community profile, what is the expected distribution of citations
from journal \code{idx}?
}
\details{
To vectorise this function, it might be easier first to create a vectorised version of
\code{\link[=nearest_point]{nearest_point()}}.
}
\examples{
distances <- as.dist(1 - cor(citations + t(citations) - diag(diag(citations))))
clusters <- cutree(hclust(distances), h = 0.6)
fitted_citations(NULL, citations, clusters)
fitted_citations('Bka', citations, clusters)

}
\seealso{
Other functions for residual analysis of communities: \code{\link{community_residuals}},
  \code{\link{profile_residuals}}
}
