% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{A Hello World Function}
\usage{
hello(friend = NULL)
}
\arguments{
\item{friend}{A character string}
}
\value{
A character greeting
}
\description{
Give me an argument and I will return you a greeting.
}
\details{
If \code{friend} is not specified, the function
  will simply say "Hello!" instead.
}
\examples{
hello()
hello('David')
hello('everyone')

}
\concept{
selby
}

