% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BradleyTerry.R
\name{hessianBT}
\alias{hessianBT}
\title{Calculate the observed Fisher information of Bradley-Terry model estimates}
\usage{
hessianBT(mu, X)
}
\description{
Calculate the observed Fisher information of Bradley-Terry model estimates
}
\seealso{
Other Bradley-Terry model utility functions: \code{\link{loglikelihood}},
  \code{\link{vcovBT}}
}

