% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILSR.R
\name{ILSR}
\alias{ILSR}
\title{Estimate the parameters of a Bradley-Terry model}
\usage{
ILSR(C, sort = FALSE, maxits = 100, tolerance = 1e-06, verbose = FALSE)
}
\arguments{
\item{C}{a square matrix of paired comparisons}

\item{sort}{logical. If \code{TRUE}, sort the weights in descending order}

\item{maxits}{The maximum number of iterations in the I-LSR algorithm}

\item{tolerance}{The criterion for convergence of the algorithm}

\item{verbose}{logical. If \code{TRUE}, show the number of iterations as a message}
}
\value{
A vector of estimated ability scores
}
\description{
Estimate Bradley-Terry ability scores from paired comparison data using the I-LSR algorithm.
Avoids using \code{glm} and seems to be faster, but only approximates maximum likelihood.
}
\details{
The function uses iterative Luce Spectral Ranking (I-LSR) to approximate the maximum likelihood ability weights of the Bradley-Terry model.
\emph{To do:} Add unit tests.
}
\examples{
ILSR(citations)

}
\references{
Maystre, Lucas and Grossglauser, Matthias (2015).
Fast and accurate inference of Plackett--Luce models.
In \emph{Advances in Neural Information Processing Systems},
172--180.
}
\seealso{
Other network centrality estimators: \code{\link{BTscores}},
  \code{\link{BradleyTerry}}, \code{\link{PageRank}},
  \code{\link{Scroogefactor}}
}

