% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiplot.R
\name{linearScale}
\alias{linearScale}
\title{Transform values on a linear scale}
\usage{
linearScale(x, codomain = c(0, 1))
}
\arguments{
\item{x}{a numeric vector}

\item{codomain}{the minimum and maximum of the output scale}
}
\value{
a vector in the same format as \code{x}, centred and scaled to fit the \code{codomain}
}
\description{
Transform values on a linear scale
}
\seealso{
Other plot constructors: \code{\link{asciiplot}},
  \code{\link{logScale}}
}

