% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTutilities.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Calculate the log-likelihood of a Bradley-Terry model}
\usage{
loglikelihood(mu, X)
}
\arguments{
\item{mu}{an estimate of ability scores}

\item{X}{a matrix of paired comparisons (weights)}
}
\description{
Calculate the log-likelihood of a Bradley-Terry model
}
\seealso{
Other Bradley-Terry model utility functions: \code{\link{BradleyTerry}},
  \code{\link{hessianBT}}, \code{\link{pseudocount}},
  \code{\link{vcovBT}}
}

