% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiplot.R
\name{logScale}
\alias{logScale}
\title{Transform values on a logarithmic scale}
\usage{
logScale(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{arguments to pass to \code{\link{linearScale}}}
}
\value{
a vector in the same format as \code{x}, centred and scaled logarithmically
}
\description{
Transform values on a logarithmic scale
}
\seealso{
Other plot.constructors: \code{\link{asciiplot}},
  \code{\link{linearScale}}
}

