% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_point.R
\name{multinomial_mix}
\alias{multinomial_mix}
\title{Multinomial mixture probabilities}
\usage{
multinomial_mix(target, profiles)
}
\arguments{
\item{target}{A vector of citations to be compared}

\item{profiles}{A matrix of \code{\link[=community_profile]{community_profile()}}s}
}
\value{
A vector of log-probabilities that each community generated \code{target}'s citation profile.
}
\description{
This is equivalent to performing the initial "E" step in expectation maximisation (EM)
for a multinomial mixture model.
}
\details{
Here we are effectively computing the nearest point in the convex hull of community profiles,
implicitly using Kullback--Leibler divergence as the distance measure.
Alternative distance measures may be used; see \code{\link[=nearest_point]{nearest_point()}} and others.
}
\examples{
# To which cluster should 'Biometrika' belong?
distances <- as.dist(1 - cor(citations + t(citations)))
clusters <- cutree(hclust(distances), h = 0.8)
profiles <- community_profile(citations, clusters)
Biometrika <- citations[, 'Bka']
w <- multinomial_mix(Biometrika, profiles)
which.max(w) == clusters['Bka']
profiles \%*\% exp(w) # nearest point

}
\seealso{
\code{\link[=nearest_point]{nearest_point()}}
}
