% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypackage.R
\docType{package}
\name{mypackage}
\alias{mypackage}
\alias{mypackage-package}
\title{Helper functions by David Selby}
\description{
A fairly random and largely unrelated assortment
  of different utilities that might come in useful.
}
\details{
The only "useful" functions you will find in here at
  the moment are \code{\link{verify}}, which is handy
  for checking simulation results, and \code{\link{PageRank}},
  a (fairly fragile!) way of scaling a matrix and computing
  its leading eigenvector.
}

