% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_point.R
\name{nearest_cosine}
\alias{nearest_cosine}
\title{Calculate nearest point by cosine similarity}
\usage{
nearest_cosine(idx, citations, communities, self = TRUE)
}
\arguments{
\item{idx}{A journal name or index.}

\item{communities}{A membership vector or \link[igraph:communities]{igraph::communities} object.}

\item{self}{logical. Include self-citations? If \code{FALSE}, they will not be counted.}
}
\value{
An object exactly like that returned by \code{\link[=nearest_point]{nearest_point()}}.
Use \code{$cosine} to extract the cosine similarity.
}
\description{
Applying EM to each individual point. Like a separate mixture model for every journal.
}
\details{
Find the nearest \link[=cprofile]{citation profile} to \code{x} that is a convex combination of
the community profiles \code{y}.

In order for an \eqn{n}-vector to be a \link[=cprofile]{citation profile},
the elements must be non-negative and sum to one.
This is also true for any convex combination (finite mixture distribution) of citation profiles.

Geometrically, this represents a point on part of the surface of the unit \eqn{n-1}-sphere
that is within the positive \href{https://en.wikipedia.org/wiki/Orthant}{closed orthant}
in \eqn{R^n}.
In a 3-journal network, this corresponds to the eighth of the unit sphere in the first octant and
in a 2-journal network, the quarter of the unit circle in the first quadrant.

The largest possible angle between two valid citation profiles is
\eqn{\theta = \frac{\pi}{2}}{\theta = \pi/2}, with cosine similarity
\eqn{\cos\theta = 0}{cos \theta = 0}.
For example, the most well-separated profiles in 2-d are \eqn{(1, 0)} and \eqn{(0, 1)} ---
i.e. the \eqn{x}-axis and the \eqn{y}-axis.
}
\seealso{
\code{\link[=nearest_point]{nearest_point()}}, \code{\link[=cosine_similarity]{cosine_similarity()}}
}
