% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageRank.R
\name{PageRank}
\alias{PageRank}
\title{Compute a PageRank vector}
\usage{
PageRank(C, alpha = 0.85, sort = FALSE)
}
\arguments{
\item{C}{a square matrix}

\item{alpha}{a damping factor}

\item{sort}{reorder the indices in descending order of PageRank}
}
\value{
A PageRank vector, scaled to sum to one
}
\description{
Given a (weighted) adjacency matrix, compute the PageRank: the
  the stationary distribution of a random walk around the graph.
}
\details{
PageRank is an eigenvector centrality metric, equivalent to Pinski & Narin's \emph{influence weight} (1976)
with the addition of a damping factor \code{alpha}, which simulates a random surfer traversing the graph and
teleporting at any time with probability \code{1 - alpha}. The effect of the damping factor is to smooth out
any disconnected components or transient portions of the network.

In bibliometrics, PageRank has also been implemented as the Eigenfactor Metric and as the SCImago Journal Rank.

By default, \code{C[i,j]} refers to the directed edge that points from column j to row i.
Use \code{t(C)} if you want edges directed from rows to columns instead.
}
\examples{
PageRank(citations)

# Scroogefactor, an estimator for the Bradley-Terry model
PageRank(citations)/colSums(citations)

}
\references{
Pinski, G., & Narin, F. (1976).
Citation influence for journal aggregates of scientific publications: Theory, with application to the literature of physics.
\emph{Information Processing & Management},
12(5), 297--312.

Page, L., Brin, S., Motwani, R., & Winograd, T. (1999).
The PageRank citation ranking: bringing order to the web.
Technical Report, Stanford InfoLab.
}

