% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTutilities.R
\name{pseudocount}
\alias{pseudocount}
\title{Add pseudocounts to a paired comparison matrix}
\usage{
pseudocount(X, alpha = 0.85)
}
\arguments{
\item{X}{a square matrix of paired comparison data}

\item{alpha}{a damping factor between 0 and 1}
}
\value{
a matrix in the same form as \code{X}, with pseudocounts added
}
\description{
Allows fitting \code{\link{BradleyTerry}} models with parameters equal to damped \code{\link{Scroogefactor}}s.
}
\details{
Eigenvector estimators can be used to compute Bradley--Terry scores.
But a \emph{damped} PageRank uses a modified probability transition matrix dependent on a parameter, \code{alpha}.
}
\examples{
SF <- Scroogefactor(citations, alpha = 0.85)
citations_pseudo <- pseudocount(citations)
BT_pseudo <- BTscores(citations_pseudo)
plot(SF, BT_pseudo); abline(0, 1)

}
\seealso{
\code{\link{BradleyTerry}}

Other Bradley-Terry model utility functions: \code{\link{BradleyTerry}},
  \code{\link{hessianBT}}, \code{\link{loglikelihood}},
  \code{\link{vcovBT}}
}

