% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multihyper.R
\name{rmultihyper}
\alias{rmultihyper}
\title{The Multinomial Hypergeometric Distribution}
\usage{
rmultihyper(n, k)
}
\arguments{
\item{n}{a vector describing the number of balls in each urn}

\item{k}{the total number of balls to draw from the urns}
}
\value{
A vector or array corresponding to the number of balls sampled from each of the urns, in the same format as \code{n}.
}
\description{
Sample a fixed number of balls, without replacement, from multiple urns.
}
\details{
Requesting a non-integer or negative number of balls, or more than are available in all the urns, will throw an error.
If \code{k} is a vector of length > 1, only the first element will be used.
Specifying only one or two urns will respectively return \code{k} or call \code{\link{rhyper}} and issue a warning.

If \eqn{k = \sum n} then the function should return \code{n}.
}
\examples{
rmultihyper(c(5, 10, 8), 6)

# 90\% sample of citation counts data
k <- round(0.9 * sum(citations))
rmultihyper(citations, k)

}
\references{
\url{http://stats.stackexchange.com/questions/199948/random-generation-from-multivariate-hypergeometric-distribution}
}
\seealso{
\code{\link{rhyper}} for sampling two different colours of balls from a single urn (equivalent to sampling one colour from two urns)
}

