% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quasi_symm.R
\name{rquasisymmetric}
\alias{rquasisymmetric}
\title{Simulate a quasi-symmetric matrix}
\usage{
rquasisymmetric(n, intensity = 10)
}
\arguments{
\item{n}{the dimension of the matrix}

\item{intensity}{the mean of the underlying Poisson random number generator}
}
\value{
A sparse, quasi-symmetric matrix of dimension \code{n} by \code{n}.
}
\description{
Simulate a quasi-symmetric matrix
}
\examples{
rquasisymmetric(5)
}

