% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrooge.R
\docType{package}
\name{scrooge}
\alias{scrooge}
\alias{scrooge-package}
\alias{scrooge-package}
\title{Helper functions by David A. Selby}
\description{
A fairly random and largely unrelated assortment
of different utilities that might come in useful.
}
\details{
The package includes \code{\link[=verify]{verify()}}, which is a handy binary operator
for checking simulation results, \code{\link[=PageRank]{PageRank()}}, a (fairly fragile!)
way of calculating PageRanks, and \code{\link[=ILSR]{ILSR()}}, an estimator for
the Bradley--Terry model. The function \code{\link[=vcovBT]{vcovBT()}} will compute asymptotic
variance-covariance matrices for the Bradley--Terry model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Selbosh/scrooge}
  \item Report bugs at \url{https://github.com/Selbosh/scrooge/issues}
}

}
\author{
\strong{Maintainer}: David Selby \email{d.selby@warwick.ac.uk}

}
