% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageRank.R
\name{Scroogefactor}
\alias{Scroogefactor}
\title{Compute the Scroogefactor}
\usage{
Scroogefactor(C, ...)
}
\arguments{
\item{C}{a square matrix}

\item{...}{other arguments to \code{\link{PageRank}}}
}
\value{
A vector equivalent to PageRank per reference
}
\description{
The Scroogefactor is PageRank divided by out-degree. It can be used as an approximate estimator for the Bradley--Terry model.
}
\details{
Pinksi and Narin (1976) first proposed this metric as a citation metric called the \emph{influence per reference}.
When applied to citation data, it can be interpreted as the influence per outgoing reference, effectively penalising larger
publications and review journals, which have larger combined bibliographies. When applied to sports or games results, it is
influence per game lost.
}
\references{
Pinski, G., & Narin, F. (1976).
Citation influence for journal aggregates of scientific publications: Theory, with application to the literature of physics.
\emph{Information Processing & Management},
12(5), 297--312.
}

