% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_selby}
\alias{theme_selby}
\title{Themes for ggplot2 and stuff}
\usage{
theme_selby(base_size = 11, base_family = "Arial Narrow")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
To make everything look a bit nicer, we opt for Arial Narrow, white backgrounds and no tick marks.
}
\examples{
\dontrun{
library(ggplot2)
ggplot(iris) +
  aes(Petal.Width, Petal.Length, colour = Species) +
  geom_point() +
  ggtitle("Example plot of Fisher's iris data",
          subtitle = "Coloured by species of iris") +
  theme_selby()
}

}
