% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BradleyTerry.R
\name{BradleyTerry}
\alias{BradleyTerry}
\title{Fit a Bradley--Terry model}
\usage{
BradleyTerry(C)
}
\arguments{
\item{C}{a square matrix of paired comparisons}
}
\value{
An object of class \code{"glm"}
}
\description{
Fits Bradley--Terry model, also known as a quasi-symmetry model,
to paired comparison data via maximum likelihood estimation.
}
\details{
This implementation avoids calling \code{\link[BradleyTerry2]{BTm}} and
instead fits the \code{\link[stats]{glm}} directly.
}
\examples{
BradleyTerry(citations)

}
\references{
Bradley, R. A., & Terry, M. E. (1952).
Rank analysis of incomplete block designs: I. The method of paired comparisons.
\emph{Biometrika},
39(3/4), 324--345.
}
\seealso{
Other Bradley-Terry model utility functions: \code{\link{hessianBT}},
  \code{\link{loglikelihood}}, \code{\link{pseudocount}},
  \code{\link{vcovBT}}

Other network centrality estimators: \code{\link{BTscores}},
  \code{\link{ILSR}}, \code{\link{PageRank}},
  \code{\link{Scroogefactor}}
}

