% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BradleyTerry.R
\name{BT2}
\alias{BT}
\alias{BT2}
\alias{BTscores}
\title{Estimate the parameters of a Bradley-Terry model}
\usage{
BT2(C, sort = FALSE)

BT(C)

BTscores(X)
}
\arguments{
\item{C}{a square matrix of paired comparisons}

\item{sort}{logical. If \code{TRUE}, sort the weights in descending order}
}
\value{
The maximum likelihood ability scores estimate, scaled to sum to one
}
\description{
Estimate Bradley-Terry ability scores from paired comparison data using an iterative scaling algorithm.
}
\details{
The current implementation, based on \code{BradleyTerry2}, is very slow and is basically a placeholder.
}
\examples{
BradleyTerry(citations)

}
\references{
Bradley, R. A., & Terry, M. E. (1952).
Rank analysis of incomplete block designs: I. The method of paired comparisons.
\emph{Biometrika},
39(3/4), 324--345.

Heather Turner, David Firth (2012).
Bradley--Terry Models in R: The \code{BradleyTerry2} Package.
\emph{Journal of Statistical Software},
48(9), 1--21.
}
\seealso{
Other Bradley-Terry model utility functions: \code{\link{hessianBT}},
  \code{\link{loglikelihood}}, \code{\link{vcovBT}}
}

