% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation-data.R
\docType{data}
\name{citations}
\alias{citations}
\alias{articles}
\title{Academic citations dataset}
\format{An object of class \code{matrix} with 47 rows and 47 columns.}
\usage{
data(citations)

data(articles)
}
\description{
A square matrix of citation counts between 47 academic journals in the field of statistics in 2010.
}
\details{
Citations point from the column journal (\code{citing}) to the row journal (\code{cited}).
The \emph{article counts} (that is, the total number of publications from that journal in 2010) are available via \code{articles}.
}
\references{
Varin, C., Cattelan, M., & Firth, D. (2016).
Statistical modelling of citation exchange between statistics journals.
\emph{Journal of the Royal Statistical Society: Series A (Statistics in Society)},
\strong{179(1)}, 1--63.
}
\seealso{
\code{\link[BradleyTerry2]{citations}}, a similar, smaller dataset in \code{BradleyTerry2}
}
\keyword{datasets}
