% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{community_residuals}
\alias{community_residuals}
\title{Calculate residuals for communities in a network}
\usage{
community_residuals(citations, communities)
}
\arguments{
\item{communities}{A membership vector or \link[igraph:communities]{igraph::communities} object.}

\item{x}{a matrix of citations (from columns to rows) or an \link[igraph:igraph]{igraph} object}
}
\description{
Calculate residuals for communities in a network
}
\seealso{
Other functions for residual analysis of communities: \code{\link{profile_residuals}}
}
