% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BradleyTerry.R
\name{hessianBT}
\alias{hessianBT}
\title{Calculate the observed Fisher information of Bradley-Terry model estimates}
\usage{
hessianBT(mu, X)
}
\arguments{
\item{mu}{a vector of Bradley--Terry model parameter estimates}

\item{X}{a square matrix of paired comparisons}
}
\description{
Calculate the observed Fisher information of Bradley-Terry model estimates
}
\seealso{
Other Bradley-Terry model utility functions: \code{\link{BT2}},
  \code{\link{loglikelihood}}, \code{\link{vcovBT}}
}

