% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PageRank.R
\name{PageRank}
\alias{PageRank}
\title{Compute a PageRank vector}
\usage{
PageRank(C, alpha = 0.85, row2col = FALSE, sort = FALSE)
}
\arguments{
\item{C}{a square matrix}

\item{alpha}{a damping factor}

\item{row2col}{by default, c_{ij} refers to the directed edge that
points from column j to row i. Set \code{TRUE} to
transpose so that edges lead from rows to columns instead}

\item{sort}{reorder the indices in descending order of PageRank}
}
\value{
A PageRank vector, scaled to sum to one
}
\description{
Given a (weighted) adjacency matrix, compute the PageRank: the
  the stationary distribution of a random walk around the graph.
}

