% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_point.R
\name{profile_residuals}
\alias{profile_residuals}
\title{Calculate Pearson profile residuals}
\usage{
profile_residuals(expected, observed)
}
\arguments{
\item{expected}{Predicted vector of citations, based on some model}

\item{observed}{Vector of observed citations from the journal of interest}
}
\value{
A numeric vector of standardised residuals, the same length as \code{expected} and \code{observed}
}
\description{
Residuals are calculated using the formula
\deqn{r_i = \frac{y_i - \hat{y}_i}{\sqrt \hat{y}_i}}{(y_i - E[y_i]) / \sqrt{E[y]_i}}
which treats each element of the citation counts vector as a Poisson variate.
}
\details{
We might want to make a generic S3 function like \code{\link[stats:residuals]{stats::residuals()}}.
\emph{Deviance} residuals may be added later. For now we have Pearson residuals only.
}
\examples{
Bka_counts <- citations[, 'Bka']
mean_profile <- rowSums(citations) / sum(citations)
mean_counts <- mean_profile * sum(Bka_counts)
profile_residuals(mean_counts, Bka_counts)

}
\references{
Agresti, Alan (2002).
\emph{Categorical Data Analysis} (2nd ed., pp. 366--367).
New York, NY: Wiley.
}
\seealso{
\code{\link[stats:rstandard.glm]{stats::rstandard.glm()}}
}
