% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrooge.R
\docType{package}
\name{scrooge}
\alias{scrooge}
\alias{scrooge-package}
\alias{scrooge-package}
\title{Helper functions by David A. Selby}
\description{
A fairly random and largely unrelated assortment
  of different utilities that might come in useful.
}
\details{
The package includes \code{\link{verify}}, which is a handy binary operator
  for checking simulation results, \code{\link{PageRank}}, a (fairly fragile!)
  way of calculating PageRanks, and \code{\link{ILSR}}, an estimator for
  the Bradley--Terry model. The function \code{\link{vcovBT}} will compute asymptotic
  variance-covariance matrices for the Bradley--Terry model.
}
