% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BradleyTerry.R
\name{vcovBT}
\alias{vcovBT}
\title{Calculate the variance-covariance matrix for a Bradley-Terry model}
\usage{
vcovBT(mu, X)
}
\arguments{
\item{mu}{a vector of Bradley--Terry model parameter estimates}

\item{X}{a square matrix of paired comparisons}
}
\value{
A Fisher information matrix
}
\description{
Calculate the variance-covariance matrix for a Bradley-Terry model
}
\examples{
# Compare results with BradleyTerry2:
if(!requireNamespace('BradleyTerry2')) install.packages('BradleyTerry2')
bincounts <- BradleyTerry2::countsToBinomial(citations)
BT2fit <- BradleyTerry2::BTm(cbind(win1, win2), player1, player2, data = bincounts)
BT2_mu <- c('AmS' = 0, coef(BT2fit)); BT2_mu <- BT2_mu - mean(BT2_mu)
BT2_covariances <- vcov(BT2fit)
my_covariances <- vcovBT(BT2_mu, citations)
plot(BT2_covariances, my_covariances); abline(0, 1)
}
\seealso{
Other Bradley-Terry model utility functions: \code{\link{hessianBT}},
  \code{\link{loglikelihood}}
}

