% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BradleyTerry.R
\name{vcovBT}
\alias{vcovBT}
\title{Calculate the variance-covariance matrix for a Bradley-Terry model}
\usage{
vcovBT(mu, X)
}
\arguments{
\item{mu}{a vector of Bradley--Terry model parameter estimates}

\item{X}{a square matrix of paired comparisons}
}
\value{
A Fisher information matrix
}
\description{
Calculate the variance-covariance matrix for a Bradley-Terry model
}
\examples{
mu <- ILSR(citations)
vcovBT(mu, citations)

}
\seealso{
Other Bradley-Terry model utility functions: \code{\link{BradleyTerry}},
  \code{\link{hessianBT}}, \code{\link{loglikelihood}}
}

