% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{\%=\%}
\alias{approx}
\alias{equalish}
\alias{isapprox}
\alias{verify}
\title{Verify if two numbers are approximately equal}
\usage{
verify(a, b)

isapprox(a, b)

a \%=\% b
}
\arguments{
\item{a}{the first vector to be compared}

\item{b}{the second vector to be compared}
}
\description{
Rounds a pair of objects to 5 decimal places
  and tests for equality, elementwise.
}
\details{
For an infix operator, use \code{a \%=\% b}.
  For more conventional evaluation, try \code{verify(a, b)}.
}
\examples{
pi \%=\% 3.141593
0.333333 \%=\% (1/3)
all(1:5 + 1e-6 \%=\% 1:5)
verify(0.7777777, 7/9)

}
\seealso{
\code{\link{all.equal}} \code{\link{Comparison}} \code{\link{identical}}
}

