% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{\%=\%}
\alias{\%=\%}
\alias{verify}
\title{Verify if two vectors are approximately equal}
\usage{
a \%=\% b

verify(a, b)
}
\arguments{
\item{a}{the first vector to be compared}

\item{b}{the second vector to be compared}
}
\description{
Returns \code{TRUE} if the vectors are about the same, or the mean relative difference if not.
}
\details{
The \code{\%=\%} function provides a quick infix (binary) version of \code{\link{all.equal}}
with the \code{tolerance} set to \code{1e-14} and \code{attributes} and \code{names} ignored.
It is useful for simulation results and other approximations that usually
won't be exactly equal on a computer using floating-point arithmetic.

Be aware that some binary operators, such as \code{`/`} take precedence,
so make sure to wrap \code{a} and \code{b} in brackets where appropriate.
Use \code{verify(a, b)} for a conventional prefix operator.
}
\examples{
0.3333333333333333 \%=\% (1/3)
verify(pi, 3.14159)
(1:5 + 1e-15) \%=\% 1:5
mean(rnorm(1e6, sd = 1e-3)) \%=\% 0

}
\seealso{
\code{\link{all.equal}} \code{\link{Comparison}} \code{\link{identical}}
}
